chore: update directory structure
