+++ Source/GameJoltAPI/Private/AsyncActions/Data-Store/FetchData.cpp
@@ -1,18 +1,18 @@
 // Copyright by Nick Lamprecht (2020-2023)
 
 
-#include "Remove.h"
+#include "AsyncActions/Data-Store/FetchData.h"
 #include "GenericPlatform/GenericPlatformHttp.h"
 
-URemove* URemove::RemoveData(EGJDataStore Scope, const FString Key)
+UFetchData* UFetchData::FetchData(EGJDataStore Scope, const FString Key)
 {
-    URemove* DataStoreNode = NewObject<URemove>();
+    UFetchData* DataStoreNode = NewObject<UFetchData>();
     DataStoreNode->Filter = Scope;
     DataStoreNode->DataKey = Key;
     return DataStoreNode;
 }
 
-void URemove::Activate()
+void UFetchData::Activate()
 {
     if(!Super::Validate())
         return;
@@ -26,16 +26,16 @@ void URemove::Activate()
     FScriptDelegate funcDelegate;
     funcDelegate.BindUFunction(this, "Callback");
 
-    FString BaseURL = "/data-store/remove/?";
-
-    BaseURL += "&key=" + FGenericPlatformHttp::UrlEncode(DataKey);
+    FString BaseURL = "/data-store/?key=" + FGenericPlatformHttp::UrlEncode(DataKey);
     FieldData = UJsonData::GetRequest(UGameJolt::CreateURL(BaseURL, Filter == EGJDataStore::user ? true : false));
     FieldData->OnGetResult.AddUnique(funcDelegate);
 }
 
-void URemove::Callback(const bool bSuccess, UJsonData* JSON)
+void UFetchData::Callback(const bool bSuccess, UJsonData* JSON)
 {
     if(!Super::VerifyResponse(bSuccess, JSON))
         return;
-    Success.Broadcast(EGJErrors::None);
+
+    FString Data = response->GetString("data");
+    Success.Broadcast(EGJErrors::None, Data);
 }
