+++ Source/GameJoltAPI/Private/AsyncActions/Data-Store/Remove.cpp
@@ -1,18 +1,18 @@
 // Copyright by Nick Lamprecht (2020-2023)
 
 
-#include "FetchData.h"
+#include "AsyncActions/Data-Store/Remove.h"
 #include "GenericPlatform/GenericPlatformHttp.h"
 
-UFetchData* UFetchData::FetchData(EGJDataStore Scope, const FString Key)
+URemove* URemove::RemoveData(EGJDataStore Scope, const FString Key)
 {
-    UFetchData* DataStoreNode = NewObject<UFetchData>();
+    URemove* DataStoreNode = NewObject<URemove>();
     DataStoreNode->Filter = Scope;
     DataStoreNode->DataKey = Key;
     return DataStoreNode;
 }
 
-void UFetchData::Activate()
+void URemove::Activate()
 {
     if(!Super::Validate())
         return;
@@ -26,16 +26,16 @@ void UFetchData::Activate()
     FScriptDelegate funcDelegate;
     funcDelegate.BindUFunction(this, "Callback");
 
-    FString BaseURL = "/data-store/?key=" + FGenericPlatformHttp::UrlEncode(DataKey);
+    FString BaseURL = "/data-store/remove/?";
+
+    BaseURL += "&key=" + FGenericPlatformHttp::UrlEncode(DataKey);
     FieldData = UJsonData::GetRequest(UGameJolt::CreateURL(BaseURL, Filter == EGJDataStore::user ? true : false));
     FieldData->OnGetResult.AddUnique(funcDelegate);
 }
 
-void UFetchData::Callback(const bool bSuccess, UJsonData* JSON)
+void URemove::Callback(const bool bSuccess, UJsonData* JSON)
 {
     if(!Super::VerifyResponse(bSuccess, JSON))
         return;
-
-    FString Data = response->GetString("data");
-    Success.Broadcast(EGJErrors::None, Data);
+    Success.Broadcast(EGJErrors::None);
 }
